Support several server instances
To do this, we can no longer rely on in-memory sharing of data: instead,
use the filesystem to store the output and status of the running
processes.

+++ compile_server/app/process_handling.py
@@ -1,3 +1,15 @@
+# This package contains two classes that can be used to launch processes
+# and process their output.
+#
+# SeparateProcess is used to launch a series of processes: for each instance,
+# a task is created to monitor the output of the processes.
+#
+# ProcessReader is used to read the current status of a process.
+#
+# The code expects that the client will make regular calls to
+# ProcessReader.poll() until the processes are completed.
+
+import os
 import shutil
 import sys
 import subprocess
@@ -17,6 +29,10 @@ class SeparateProcess(object):
         self.cmd_lines = cmd_lines
         self.q = Queue()
         self.working_dir = cwd
+        self.output_file = os.path.join(self.working_dir, 'output.txt')
+        self.status_file = os.path.join(self.working_dir, 'status.txt')
+        with open(self.status_file, 'wb') as f:
+            f.write("")
         self.p = None
         t = Thread(target=self._enqueue_output)
         t.daemon = True
@@ -35,32 +51,69 @@ class SeparateProcess(object):
                 bufsize=1,
                 close_fds=True)
 
+            # Write the output line by line in the output file
             for line in iter(self.p.stdout.readline, b''):
-                self.q.put(line)
+                with open(self.output_file, 'ab') as f:
+                    f.write(line)
 
+            # Write the return code
+            self.p.poll()
+            returncode = self.p.returncode
+            with open(self.status_file, 'wb') as f:
+                f.write(str(returncode))
+
+            # Cleanup
             self.p.stdout.close()
 
-        # When all the processes are complete,
-        shutil.rmtree(self.working_dir)
+            # If the process returned nonzero, do not run the next process
+            if returncode != 0:
+                break
+
+
+class ProcessReader(object):
+
+    def __init__(self, working_dir):
+        self.working_dir = working_dir
+        self.output_file = os.path.join(self.working_dir, 'output.txt')
+        self.status_file = os.path.join(self.working_dir, 'status.txt')
 
     def poll(self):
         """ Check whether the process is still running.
             return None if the process is still running, otherwise return
             the status code.
         """
-        if not self.p:
-            return None
-        self.p.poll()
-        return self.p.returncode
+        # Custom debug codes, to debug application
+        if not os.path.isdir(self.working_dir):
+            return 101
 
-    def read_lines(self):
+        if not os.path.isfile(self.status_file):
+            return None
+
+        with open(self.status_file) as f:
+            status_text = f.read().strip()
+
+        if not status_text:
+            return None
+        else:
+            # When all the processes are completed, remove the working dir
+            shutil.rmtree(self.working_dir)
+            return int(status_text)
+
+    def read_lines(self, already_read=0):
         """Read all the available lines from the process.
-           Return an empty list if there"""
-        lines = []
-        while True:
-            try:
-                # Read the queue
-                line = self.q.get_nowait()
-                lines.append(line)
-            except Empty:
-                return lines
+           already_read indicates the number of lines that have already been
+           read by the process.
+           Return an empty list if there is nothing to read.
+        """
+
+        # Custom debug codes, to debug application
+        if not os.path.isdir(self.working_dir):
+            return 103
+
+        if not os.path.isfile(self.output_file):
+            return []
+
+        with open(self.output_file, "rb") as f:
+            lines = f.readlines()
+
+        return lines[already_read:]
